Removed PropertyDialog Abstraction
